changed no_profiles and profile type from int to uint8_t
- no_profiles >= 1
- profile > 0
- added vscode to gitignore

+++ include/nigiri/loader/build_footpaths.h
@@ -4,6 +4,6 @@
 
 namespace nigiri::loader {
 
-void build_footpaths(timetable& tt, int const& no_profiles);
+void build_footpaths(timetable& tt, uint8_t const& no_profiles);
 
 }  // namespace nigiri::loader
